+++ index.ts
@@ -1,7 +1,6 @@
 import express, { Express } from 'express';
 
-import authRoutes from './routes/auth.route';
-import postRoutes from './routes/post.route';
+import routes from './routes/app.route';
 import config from './config/app.config';
 import dbConnect from './utils/databaseConnect';
 import cors from './middlewares/cors.middleware';
@@ -13,8 +12,7 @@ app.use(cors);
 
 dbConnect(config.db.MONGODB_URI);
 
-app.use('/auth', authRoutes);
-app.use('/post', postRoutes);
+app.use('/api', routes);
 
 app.listen(config.app.PORT, () => {
 	console.log(`App listening on port ${config.app.PORT}`);