+++ routes/vote.route.ts
@@ -3,3 +3,5 @@ import authorize from '../middlewares/authorize.middleware';
 import vote from '../controllers/vote.controller';
 
 router.post('/add', authorize, vote.addVote);
+
+export default router;